# docs/openapi/critical-stable.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DashClaw Critical Stable API",
    "version": "4.45.0",
    "description": "Generated OpenAPI spec for critical stable DashClaw endpoints. Source: app/api/**/route.js"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/guard": {
      "get": {
        "operationId": "get_api_guard",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or request error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "GET /api/guard",
        "tags": [
          "guard"
        ],
        "x-api-maturity": "stable"
      },
      "post": {
        "operationId": "post_api_guard",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Validation or request error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "POST /api/guard",
        "tags": [
          "guard"
        ],
        "x-api-maturity": "stable"
      }
    },
...
  }
}